almacen/requirements.txt
Flask
Flask-SQLAlchemy
pyyaml